Simplify lwjgl rule
