+++ buildSrc/src/main/kotlin/org/openrndr/convention/LwjglRule.gradle.kts
@@ -6,8 +6,10 @@ abstract class LwjglRule : ComponentMetadataRule {
     @get:Inject
     abstract val objects: ObjectFactory
 
-    override fun execute(context: ComponentMetadataContext) {
-        context.details.withVariant("runtime") {
+    override fun execute(context: ComponentMetadataContext) = context.details.run {
+        if (id.group != "org.lwjgl") return
+        if (id.name == "lwjgl-egl") return
+        withVariant("runtime") {
             attributes {
                 attributes.attribute(OperatingSystemFamily.OPERATING_SYSTEM_ATTRIBUTE, objects.named("none"))
                 attributes.attribute(MachineArchitecture.ARCHITECTURE_ATTRIBUTE, objects.named("none"))
@@ -15,13 +17,13 @@ abstract class LwjglRule : ComponentMetadataRule {
         }
         for (nativeVariant in openrndrJvmNativeVariants) {
             val lwjglTargetName = nativeVariant.mapToLwjglTargetName()
-            context.details.addVariant("$lwjglTargetName-runtime", "runtime") {
+            addVariant("$lwjglTargetName-runtime", "runtime") {
                 attributes {
                     attributes.attribute(OperatingSystemFamily.OPERATING_SYSTEM_ATTRIBUTE, objects.named(nativeVariant.os))
                     attributes.attribute(MachineArchitecture.ARCHITECTURE_ATTRIBUTE, objects.named(nativeVariant.arch))
                 }
                 withFiles {
-                    addFile("${context.details.id.name}-${context.details.id.version}-$lwjglTargetName.jar")
+                    addFile("${id.name}-${id.version}-$lwjglTargetName.jar")
                 }
             }
